+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - OpenRouter provider
+### Changed
+- Test system modified
 ## [0.2] - 2025-12-05
 ### Added
 - Logo