+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - OpenRouter provider
 ### Changed
 - Test system modified
+- `README.md` updated
 ## [0.2] - 2025-12-05
 ### Added
 - Logo