+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.3] - 2025-12-17
 ### Added
 - OpenRouter provider
 - Cerebras provider
@@ -33,7 +34,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `--mode` argument
 - `--tone` argument
 
-[Unreleased]: https://github.com/sepandhaghighi/mytext/compare/v0.2...dev
+[Unreleased]: https://github.com/sepandhaghighi/mytext/compare/v0.3...dev
+[0.3]: https://github.com/sepandhaghighi/mytext/compare/v0.2...v0.3
 [0.2]: https://github.com/sepandhaghighi/mytext/compare/v0.1...v0.2
 [0.1]: https://github.com/sepandhaghighi/mytext/compare/dde63ee...v0.1